+++ service/config.py
@@ -19,7 +19,7 @@ if 'VCAP_SERVICES' in os.environ:
 # Configure SQLAlchemy
 SQLALCHEMY_DATABASE_URI = DATABASE_URI
 SQLALCHEMY_TRACK_MODIFICATIONS = False
-SQLALCHEMY_POOL_SIZE = 2
+#SQLALCHEMY_POOL_SIZE = 2
 
 # Secret for session management
 SECRET_KEY = os.getenv("SECRET_KEY", "sup3r-s3cr3t")
