+++ tests/test_cli_commands.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from unittest.mock import patch, MagicMock
 from click.testing import CliRunner
-from service.utils.cli_commands import create_db
+from service.common.cli_commands import create_db
 
 
 class TestFlaskCLI(TestCase):
@@ -10,7 +10,7 @@ class TestFlaskCLI(TestCase):
     def setUp(self):
         self.runner = CliRunner()
 
-    @patch('service.utils.cli_commands.db')
+    @patch('service.common.cli_commands.db')
     def test_create_db(self, db_mock):
         """It should call the create-db command"""
         db_mock.return_value = MagicMock()
